Version 0.77
This is a minor bug fix version of the previous version.  Namely, the dual
simplex method might cycle in the Phase I, although it is rather rare.   The
fix is for the options "vertex_listing" and "facet_listing" which run the dual
simplex as a default but switches to the criss-cross method when it generates
too many pivots.  The cycling examples were reported by Boryeu Mao
(bmao@Raven.pnu.com).  This version can be compiled with gcc-3.1.* or
gcc-3.2.*.  The earlier version gcc-2.95.* is not compatible with this release.

Imported in Apr 2018 from cdd+-077.tar.gz released Aug 19, 2003.